renamed to ignore old stuff, 'fix' is now new bookies.js
