+++ front-end/src/tests/bookies/bookies.test.js
@@ -752,7 +752,7 @@ const bigBookiesInflated = [
 
 
 import { describe, expect, test, beforeEach } from '@jest/globals';
-import { Rebuild, FlatParent } from '../../modules/bookies/bookies.fix';
+import { Rebuild, FlatParent } from '../../modules/bookies/bookies';
 
 
 describe("[FlatParent] Rebuilds a folder-item into its original shape.", () => {
